rest_api_service/requirements.txt
requests
psycopg2-binary
flask
Flask-SQLAlchemy
python-Levenshtein
